lucas

+++ ocr_folder/Packages/lucas/Binary.c
@@ -1,3 +1,6 @@
+#include <string.h>
+#include "../Thomas/TraitementImage.h"
+#include <math.h>
 #include "Binary.h"
 
 
@@ -237,8 +240,8 @@ BMPPic_ Changing (BMPPic_ pic,int threshold,size_t x,size_t y)
 //aplly the binarisation
 BMPPic_ end (BMPPic_ pic)
 {
-    size_t i = 0;
-    size_t j = 0;
+    int i = 0;
+    int j = 0;
     histogram histo = Get_histo(pic);
     histo = Get_window(histo);
     histo = Calculus(histo);
